Add files via upload
This script copies files to a new folder after extracting them from the original folder.
The initial step in creating training data is to find suitable frames and
make masks. Because the use of the same side and bottom frames is preferable for
training data. The names of the side and bottom frames must match.
If you made masks from the side frames, 
you can move the png files from the original bottom folder to a new folder 
by using names/frames used in side frames.